requirements.txt
tensorflow==2.13.0
tf-agents==0.17.0
pandas
yfinance
pyyaml
scikit-learn